requirements.txt
black==24.4.2
click==8.1.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-constraint==1.4.0
tomli==2.0.1
typing_extensions==4.12.2
